fix interactive TUI
it wasn't able to cope with the new 'init' vertex that we create while
initializing

along the way, also remove the redundant 'Duration' status info now that
Progrock displays timing info natively in its UI
